LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/text/2d.text.drawing.style.textRendering.settings.worker-expected.txt


FAIL Testing basic functionalities of textRendering in Canvas assert_equals: ctx.textRendering === "auto" (got [undefined], expected auto[string]) expected (string) "auto" but got (undefined) undefined
